feat: Implement card reveal on click
Changed cards state object so that class names can be determined
by Game component. Added more props to Card to handle class
updates and callback for onClick event. Right now clicking a
card will reveal it and start the timer if it's the first card.
Need to look into hooks for these sorts of updates more to see
if it's more intuitive to use a hook instead of simply calling
a setState function.