fix: Prevent click during bad match, correctly reset
Previously, the original game's state was altered when setting
state. Now, a copy is created of the original state when applicable
and alters that copy before setting the state. Additionally,
a user could click other cards when a bad match animation was
going which could lead to scenarios where the subsequent match
did not show the next two cards, but just the first. The onClick
event is now ignored temporarily while the bad match is shown.